moving files and improving code
