+++ packages/app/src/components/inspector/FillRow.tsx
@@ -81,7 +81,7 @@ export default memo(function FillRow({
         <Spacer.Horizontal size={8} />
         <InputField.Root id={hexInputId} labelPosition="start">
           <InputField.Input
-            value={`${sketchColorToHex(color).replace('#', '').toUpperCase()}`}
+            value={`${sketchColorToHex(color).replace('#', '')}`}
             onSubmit={() => {}}
           />
           <InputField.Label>#</InputField.Label>
